Updating projects and formatting for sub-priorities.
The formatting I had for the sub-priorities wasn't in the form of a
sub-list, so that's been addressed (and verified via pandoc).

I've updated the existing projects that are currently in progress and
added a few that are on deck (and a few new issues have been created in
the softwareteam repo for them).